Add schema folder
